Adds a debugging option
The point here is that there are cases where the server decides that the request
(either preflight or actual) is not correct and doesn't respond with the correct
headers.  It is tricky to debug these cases.  So this adds an option to turn on
a log that explains the actions of the CORS middleware.