Added support for "*" in the specification of AllowedHeaders
These changes treat `"*"` as a special case (that matches any header) within the
`AllowedHeaders` array.  A test case was added to check the functionality of
this wildcard and the documentation regarding the `AllowedHeaders` field was
updated to mention this functionality as well as the automatic appending
of the `"Origin"` header (i.e., that it is not necessary to specify the
`"Origin"` header among the set of allowed headers).