Update README.md for ease of use
Added Debug to one of the examples, explained why with a comment.
I know the information is there but I don't think it is well highlighted. I was wondering why my requests were coming back without Access-Control-Allow-Origin (but not erroring visibly) so perhaps this will help others not make the same errors.